feat: Add updateUser and deleteUser functions
The code changes in this commit add two new functions, `updateUser` and `deleteUser`, to the `users.js` file in the `service` directory. These functions handle updating and deleting user data respectively. They make network requests to the server using the `fetch` API and include the necessary headers for authentication. If the request is successful, the updated user data is returned. If there is an error, an error message is logged and the error is thrown.

The commit message follows the conventional format of starting with a type (feat) followed by a brief description of the changes.